backend/requirements.txt
flask
mysql-connector-python
pdfplumber
python-dotenv
flask-cors